requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.32
psycopg2-binary==2.9.9

python-dotenv==1.0.1

passlib==1.7.4
bcrypt==3.2.2

python-jose==3.3.0

pydantic==2.9.2
email-validator==2.2.0

python-multipart==0.0.9

supabase>=2.5.0

numpy
opencv-python
requests
rapidfuzz
anthropic
pillow-heif
Pillow
paddleocr
paddlepaddle

# VLM-based OCR engines (MiniCPM-V + PaddleOCR-VL)
transformers>=5.0.0
sentencepiece          # MiniCPM-V tokenizer
timm                   # MiniCPM-V vision encoder (SigLIP)
bitsandbytes           # 4-bit NF4 quantization for MiniCPM-V on <22 GB VRAM GPUs (T4 16 GB, RTX 4060 8 GB, etc.)
qwen-vl-utils          # PaddleOCR-VL image pre-processing
num2words              # required by SmolVLM2 processor
# CUDA build — install with: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
# For Render (CPU-only), remove the index-url and use plain: torch torchvision
torch
torchvision
einops
accelerate